+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Feat: Automatically enable `Timber` and `Fragment` integrations if they are present on the classpath (#1936)
+
 ## 5.6.3
 
 * Fix: If transaction or span is finished, do not allow to mutate (#1940)